refactor: break circular dependency workspace-layout ↔ workspace-serializer
Extract disposeTab/disposeAllTabs into workspace-cleanup.js so that
workspace-serializer.js no longer imports from workspace-layout.js.
workspace-layout.js re-exports for backward compat.

Closes #94
