+++ src/utils/workspace-serializer.js
@@ -21,7 +21,7 @@ import { generateId } from './id.js';
 import { WorkspaceTab } from './tab-manager-helpers.js';
 import { disposeAllSideViews } from './sidebar-manager.js';
 import { capturePanelWidths } from './workspace-resize.js';
-import { disposeAllTabs } from './workspace-layout.js';
+import { disposeAllTabs } from './workspace-cleanup.js';
 
 // ── Serialization ──
